Application completed for queues and stacks
full application with basic GUI completed for demonstration using queues and stacks. The GUI allows you to: Input data into a queue, remove data from the back of the queue, swap the queue order using a stack, print the contents of the queue, clear the entire queue of its contents and exit application.